Fixed Agent architecture
